requirements.txt
pytest
selenium
webdriver-manager
appium-python-client